Make the header file completely private
