+++ opencog/persist/gearman/CMakeLists.txt
@@ -5,10 +5,6 @@ ADD_LIBRARY(dist-gearman
 	DistSCM.cc
 )
 
-ADD_DEPENDENCIES(dist-gearman
-	opencog_atom_types
-)
-
 TARGET_LINK_LIBRARIES(dist-gearman
 	smob
 	atomspace
@@ -16,7 +12,3 @@ TARGET_LINK_LIBRARIES(dist-gearman
 )
 
 INSTALL (TARGETS dist-gearman DESTINATION "lib${LIB_DIR_SUFFIX}/opencog")
-INSTALL (FILES
-	DistSCM.h
-	DESTINATION "include/opencog/persist/gearman"
-)
